mcp-server/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
httpx==0.25.0
python-dotenv==1.0.0
openai==1.12.0
pyyaml==6.0        # For parsing OpenAPI YAML
boto3==1.28.57